requirements.txt
streamlit
python-dotenv
pandas
PyMuPDF
sentence-transformers
google-generativeai
openpyxl
